etl/requirements.txt
langchain
openai
langchain-openai
langchain-pinecone
beautifulsoup4
requests
tqdm
langchain-community
pinecone-client
pydub
tqdm
requests
pytube
pytubefix
ffmpeg-python
llama-index-core
llama-index-llms-openai
llama-index-vector-stores-pinecone